requirements.txt
temporalio>=1.5.0
google-generativeai>=0.3.0
openai>=1.1.0
anthropic>=0.18.0
python-dotenv>=1.0.0
pydantic>=2.4.0
flask>=2.3.0
werkzeug>=2.3.0
gunicorn>=21.2.0
requests>=2.31.0
Pillow>=10.0.0  # For handling images easily